refactor(wave5-step3): rename verify internals and move tests off facade
