+++ crates/assay-registry/src/verify_internal/digest.rs
@@ -5,7 +5,9 @@
 //! - no DSSE signature verification
 //! - no policy decisions
 
-use crate::canonicalize::{compute_canonical_digest, CanonicalizeError};
+use crate::canonicalize::{
+    compute_canonical_digest, parse_yaml_strict, to_canonical_jcs_bytes, CanonicalizeError,
+};
 use crate::digest::{compute_canonical_or_raw_digest, sha256_hex_bytes};
 use crate::error::{RegistryError, RegistryResult};
 
@@ -37,4 +39,14 @@ pub(crate) fn compute_digest_strict_impl(content: &str) -> Result<String, Canoni
 
 pub(crate) fn compute_digest_raw_impl(content: &str) -> String {
     sha256_hex_bytes(content.as_bytes())
+}
+
+pub(crate) fn canonicalize_for_dsse_impl(content: &str) -> RegistryResult<Vec<u8>> {
+    let json_value = parse_yaml_strict(content).map_err(|e| RegistryError::InvalidResponse {
+        message: format!("failed to parse YAML for signature verification: {}", e),
+    })?;
+
+    to_canonical_jcs_bytes(&json_value).map_err(|e| RegistryError::InvalidResponse {
+        message: format!("failed to canonicalize for signature verification: {}", e),
+    })
 }
