+++ crates/assay-registry/src/verify_internal/errors.rs
