+++ crates/assay-registry/src/verify_internal/keys.rs
