+++ crates/assay-registry/src/verify_internal/mod.rs
@@ -1,0 +1,13 @@
+//! Internal verify implementation modules.
+//!
+//! `crate::verify` is the permanent public facade. This module owns the split
+//! implementation boundaries and test module for Wave5 Step3 closure.
+
+pub(crate) mod digest;
+pub(crate) mod dsse;
+pub(crate) mod errors;
+pub(crate) mod keys;
+pub(crate) mod policy;
+#[cfg(test)]
+pub(crate) mod tests;
+pub(crate) mod wire;
