+++ crates/assay-registry/src/verify_internal/wire.rs
